+++ internal/database/controller/database.go
@@ -2,12 +2,8 @@ package database
 
 import (
 	"context"
-	"errors"
 	mod "go-cdn/pkg/model"
 )
-
-var ErrDatabaseOp = errors.New("error on database operation")
-var ErrKeyDoesNotExist = errors.New("key does not exist")
 
 type databaseRepository interface {
 	GetFile(ctx context.Context, id_hash_search string) (*mod.StoredFile, error)
