+++ internal/discovery/repository/dummy/dummy.go
@@ -1,4 +1,6 @@
-package discovery
+package dummy
+
+import "go-cdn/internal/discovery/repository"
 
 type DummyRepository struct{}
 
@@ -7,9 +9,9 @@ func NewDummyRepo() *DummyRepository {
 	return &DummyRepository{}
 }
 
-func (d *DummyRepository) RegisterService() error { return nil }
+func (d *DummyRepository) RegisterService() error { return repository.ErrServiceDisabled }
 
-func (d *DummyRepository) DeregisterService() error { return nil }
+func (d *DummyRepository) DeregisterService() error { return repository.ErrServiceDisabled }
 
 func (d *DummyRepository) DiscoverService(address string) ([]string, error) {
 	return []string{address}, nil
