+++ internal/discovery/repository/error.go
@@ -1,0 +1,6 @@
+package repository
+
+import "errors"
+
+var ErrServiceNotFound = errors.New("service not found")
+var ErrServiceDisabled = errors.New("this discovery service is disabled")
